+++ crates/miden-objects/src/account/delta/mod.rs
@@ -18,7 +18,8 @@ pub use vault::{
 // ACCOUNT DELTA
 // ================================================================================================
 
-/// [AccountDelta] stores the differences between two account states.
+/// The [`AccountDelta`] stores the differences between two account states, which can result from
+/// one or more transaction.
 ///
 /// The differences are represented as follows:
 /// - storage: an [AccountStorageDelta] that contains the changes to the account storage.
@@ -287,8 +288,18 @@ impl SequentialCommit for AccountDelta {
 // ACCOUNT UPDATE DETAILS
 // ================================================================================================
 
-/// Describes the details of an account state transition resulting from applying a transaction to
-/// the account.
+/// [`AccountUpdateDetails`] describes the details of one or more transactions executed against an
+/// account.
+///
+/// In particular, private account changes aren't tracked at all; they are represented as
+/// [`AccountUpdateDetails::Private`].
+///
+/// New non-private accounts are included in full and changes to a non-private account are tracked
+/// as an [`AccountDelta`].
+///
+/// Note that these details can represent the changes from one or more transactions in which case
+/// the delta is either applied to the new account or deltas are merged together using
+/// [`AccountDelta::merge`].
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum AccountUpdateDetails {
     /// Account is private (no on-chain state change).
@@ -440,7 +451,7 @@ fn validate_nonce(
     vault: &AccountVaultDelta,
 ) -> Result<(), AccountDeltaError> {
     if (!storage.is_empty() || !vault.is_empty()) && nonce_delta == ZERO {
-        return Err(AccountDeltaError::ZeroNonceForNonEmptyDelta);
+        return Err(AccountDeltaError::NonEmptyStorageOrVaultDeltaWithZeroNonceDelta);
     }
 
     Ok(())
@@ -485,7 +496,7 @@ mod tests {
         assert_matches!(
             AccountDelta::new(account_id, storage_delta.clone(), vault_delta.clone(), ZERO)
                 .unwrap_err(),
-            AccountDeltaError::ZeroNonceForNonEmptyDelta
+            AccountDeltaError::NonEmptyStorageOrVaultDeltaWithZeroNonceDelta
         );
         AccountDelta::new(account_id, storage_delta.clone(), vault_delta.clone(), ONE).unwrap();
     }