+++ crates/miden-objects/src/errors.rs
@@ -251,8 +251,8 @@ pub enum AccountDeltaError {
         account_id: AccountId,
         source: AccountError,
     },
-    #[error("zero nonce is not allowed for non-empty account deltas")]
-    ZeroNonceForNonEmptyDelta,
+    #[error("non-empty account storage or vault delta with zero nonce delta is not allowed")]
+    NonEmptyStorageOrVaultDeltaWithZeroNonceDelta,
     #[error(
         "account nonce increment {current} plus the other nonce increment {increment} overflows a felt to {new}"
     )]